requirements.txt
fastapi[standard]
jinja2
uvicorn
pandas
sqlalchemy
pymysql
pydantic
openpyxl
pyjwt